Remove separation between `user` and `reservedemail` java packages.
